Refactor notification request content record
